tiny compiler warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@6325 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtui/InstanceDlg.cpp
@@ -83,7 +83,7 @@ vtTagArray *InstanceDlg::GetTagArray()
 		vtContentManager *cman = Current();
 		if (!cman)
 			return NULL;
-		if (m_iItem >= cman->NumItems())
+		if (m_iItem >= (int) cman->NumItems())
 			return NULL;
 		vtItem *item = cman->GetItem(m_iItem);
 		if (!item)
